+++ common/persistence/nosql/nosqlplugin/cassandra/gocql/public/testBase.go
@@ -19,10 +19,9 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package tests
+package public
 
 import (
-	_ "github.com/uber/cadence/common/persistence/nosql/nosqlplugin/cassandra/gocql/public" // needed to load the default gocql client
 	persistencetests "github.com/uber/cadence/common/persistence/persistence-tests"
 )
