+++ common/persistence/nosql/nosqlplugin/cassandra/tests/cassandra_test.go
@@ -25,61 +25,62 @@ import (
 
 	"github.com/stretchr/testify/suite"
 
+	"github.com/uber/cadence/common/persistence/nosql/nosqlplugin/cassandra/gocql/public"
 	persistencetests "github.com/uber/cadence/common/persistence/persistence-tests"
 )
 
 func TestCassandraHistoryV2Persistence(t *testing.T) {
 	s := new(persistencetests.HistoryV2PersistenceSuite)
-	s.TestBase = NewTestBaseWithPublicCassandra(&persistencetests.TestBaseOptions{})
+	s.TestBase = public.NewTestBaseWithPublicCassandra(&persistencetests.TestBaseOptions{})
 	s.TestBase.Setup()
 	suite.Run(t, s)
 }
 
 func TestCassandraMatchingPersistence(t *testing.T) {
 	s := new(persistencetests.MatchingPersistenceSuite)
-	s.TestBase = NewTestBaseWithPublicCassandra(&persistencetests.TestBaseOptions{})
+	s.TestBase = public.NewTestBaseWithPublicCassandra(&persistencetests.TestBaseOptions{})
 	s.TestBase.Setup()
 	suite.Run(t, s)
 }
 
 func TestCassandraMetadataPersistenceV2(t *testing.T) {
 	s := new(persistencetests.MetadataPersistenceSuiteV2)
-	s.TestBase = NewTestBaseWithPublicCassandra(&persistencetests.TestBaseOptions{})
+	s.TestBase = public.NewTestBaseWithPublicCassandra(&persistencetests.TestBaseOptions{})
 	s.TestBase.Setup()
 	suite.Run(t, s)
 }
 
 func TestCassandraShardPersistence(t *testing.T) {
 	s := new(persistencetests.ShardPersistenceSuite)
-	s.TestBase = NewTestBaseWithPublicCassandra(&persistencetests.TestBaseOptions{})
+	s.TestBase = public.NewTestBaseWithPublicCassandra(&persistencetests.TestBaseOptions{})
 	s.TestBase.Setup()
 	suite.Run(t, s)
 }
 
 func TestCassandraVisibilityPersistence(t *testing.T) {
 	s := new(persistencetests.VisibilityPersistenceSuite)
-	s.TestBase = NewTestBaseWithPublicCassandra(&persistencetests.TestBaseOptions{})
+	s.TestBase = public.NewTestBaseWithPublicCassandra(&persistencetests.TestBaseOptions{})
 	s.TestBase.Setup()
 	suite.Run(t, s)
 }
 
 func TestCassandraExecutionManager(t *testing.T) {
 	s := new(persistencetests.ExecutionManagerSuite)
-	s.TestBase = NewTestBaseWithPublicCassandra(&persistencetests.TestBaseOptions{})
+	s.TestBase = public.NewTestBaseWithPublicCassandra(&persistencetests.TestBaseOptions{})
 	s.TestBase.Setup()
 	suite.Run(t, s)
 }
 
 func TestCassandraExecutionManagerWithEventsV2(t *testing.T) {
 	s := new(persistencetests.ExecutionManagerSuiteForEventsV2)
-	s.TestBase = NewTestBaseWithPublicCassandra(&persistencetests.TestBaseOptions{})
+	s.TestBase = public.NewTestBaseWithPublicCassandra(&persistencetests.TestBaseOptions{})
 	s.TestBase.Setup()
 	suite.Run(t, s)
 }
 
 func TestQueuePersistence(t *testing.T) {
 	s := new(persistencetests.QueuePersistenceSuite)
-	s.TestBase = NewTestBaseWithPublicCassandra(&persistencetests.TestBaseOptions{})
+	s.TestBase = public.NewTestBaseWithPublicCassandra(&persistencetests.TestBaseOptions{})
 	s.TestBase.Setup()
 	suite.Run(t, s)
 }
